Move additional compatibility code to the `compat` module
